+++ models/demos/wormhole/stable_diffusion/tt/ttnn_functional_resnetblock2d_new_conv.py
@@ -432,6 +432,8 @@ class resnetBlock2D:
                 input_width=self.conv1_input_width,
                 conv_config=conv_config,
                 conv_op_cache=conv_cache,
+                return_output_size=True,
+                return_prepared_device_weights=True,
             )
 
         else:
@@ -510,6 +512,8 @@ class resnetBlock2D:
                     input_width=self.conv1_input_width,
                     conv_config=conv_config,
                     conv_op_cache=conv_cache,
+                    return_output_size=True,
+                    return_prepared_device_weights=True,
                 )
                 if i != 0:
                     split_hidden_states[i] = ttnn.add(
@@ -630,6 +634,8 @@ class resnetBlock2D:
             input_width=self.conv2_input_width,
             conv_config=conv_config,
             conv_op_cache=conv_cache,
+            return_output_size=True,
+            return_prepared_device_weights=True,
         )
         use_in_shortcut = in_channels != out_channels if use_in_shortcut is None else use_in_shortcut
 
@@ -672,6 +678,8 @@ class resnetBlock2D:
                 input_width=self.conv_shortcut_input_width,
                 conv_config=conv_config,
                 conv_op_cache=conv_cache,
+                return_output_size=True,
+                return_prepared_device_weights=True,
             )
 
         if ttnn.get_memory_config(input_tensor) != ttnn.get_memory_config(hidden_states):
