+++ models/demos/wormhole/stable_diffusion/tt/ttnn_functional_upsample_2d_new_conv.py
@@ -118,5 +118,7 @@ class upsample2d:
             bias_tensor=self.conv_bias_tensor,
             conv_config=conv_config,
             conv_op_cache=conv_cache,
+            return_output_size=True,
+            return_prepared_device_weights=True,
         )
         return tt_out
