+++ models/demos/yolov4/ttnn/common.py
@@ -90,7 +90,7 @@ class Conv:
         if self.act_block_h is not None:
             conv_config.act_block_h_override = self.act_block_h
 
-        [output_tensor, _out_height, _out_width, self.weights, self.bias] = ttnn.conv2d(
+        [output_tensor, self.weights, self.bias] = ttnn.conv2d(
             input_tensor=input_tensor,
             weight_tensor=self.weights,
             bias_tensor=self.bias,
@@ -104,5 +104,6 @@ class Conv:
             input_height=self.input_params[1],
             input_width=self.input_params[2],
             conv_config=conv_config,
+            return_prepared_device_weights=True,
         )
         return output_tensor
