+++ models/experimental/functional_unet/tt/unet_shallow_ttnn.py
@@ -194,7 +194,7 @@ class UNetConv2D:
         self.bias = ttnn.from_torch(bias, dtype=ttnn.float32, mesh_mapper=mesh_mapper)
 
     def __call__(self, x):
-        x, _, _, self.weight, self.bias = ttnn.conv2d(
+        x, self.weight, self.bias = ttnn.conv2d(
             input_tensor=x,
             weight_tensor=self.weight,
             bias_tensor=self.bias,
@@ -209,6 +209,7 @@ class UNetConv2D:
             padding=self.padding,
             conv_config=self.conv_config,
             conv_op_cache=self.cache,
+            return_prepared_device_weights=True,
         )
         return x
