+++ tests/ttnn/unit_tests/operations/test_new_conv2d.py
@@ -168,6 +168,8 @@ def run_conv(
         debug=debug,
         groups=groups,
         memory_config=memory_config,
+        return_output_size=True,
+        return_prepared_device_weights=True,
     )
 
     tt_output_tensor = ttnn.from_device(tt_output_tensor_on_device)
@@ -306,6 +308,8 @@ def run_conv_with_split(
             input_width=input_width,
             conv_config=conv_config,
             conv_op_cache=reader_patterns_cache,
+            return_output_size=True,
+            return_prepared_device_weights=True,
         )
         tt_conv_output_tensor = ttnn.from_device(tt_output_tensor_on_device)
         torch_conv_output_tensor = ttnn.to_torch(tt_conv_output_tensor)
@@ -561,6 +565,8 @@ def test_conv_ws(
         conv_op_cache=reader_patterns_cache,
         debug=debug,
         groups=groups,
+        return_output_size=True,
+        return_prepared_device_weights=True,
     )
 
     tt_output_tensor = ttnn.from_device(tt_output_tensor_on_device)
