+++ ttnn/cpp/ttnn/operations/conv/conv2d/conv2d_pybind.cpp
@@ -9,6 +9,9 @@
 #include "conv2d_pybind.hpp"
 #include "ttnn/cpp/ttnn/operations/sliding_window/sliding_window_pybind.hpp"
 #include "conv2d.hpp"
+#include "conv2d_utils.hpp"
+#include "prepare_conv2d_weights.hpp"
+#include "ttnn/types.hpp"
 
 namespace py = pybind11;
 
@@ -121,6 +124,85 @@ void py_bind_conv2d(py::module& module) {
     );
 
     module.def(
+        "prepare_conv_weights",
+        prepare_conv_weights<ttnn::Device>,
+        py::kw_only(),
+        py::arg("weight_tensor"),
+        py::arg("input_memory_config"),
+        py::arg("weights_format"),
+        py::arg("in_channels"),
+        py::arg("out_channels"),
+        py::arg("batch_size"),
+        py::arg("input_height"),
+        py::arg("input_width"),
+        py::arg("kernel_size"),
+        py::arg("stride"),
+        py::arg("padding"),
+        py::arg("dilation"),
+        py::arg("groups"),
+        py::arg("device"),
+        py::arg("conv_config") = std::nullopt);
+
+
+    module.def(
+        "prepare_conv_weights",
+        prepare_conv_weights<ttnn::MeshDevice>,
+        py::kw_only(),
+        py::arg("weight_tensor"),
+        py::arg("input_memory_config"),
+        py::arg("weights_format"),
+        py::arg("in_channels"),
+        py::arg("out_channels"),
+        py::arg("batch_size"),
+        py::arg("input_height"),
+        py::arg("input_width"),
+        py::arg("kernel_size"),
+        py::arg("stride"),
+        py::arg("padding"),
+        py::arg("dilation"),
+        py::arg("groups"),
+        py::arg("device"),
+        py::arg("conv_config") = std::nullopt);
+
+    module.def(
+        "prepare_conv_bias",
+        prepare_conv_bias<ttnn::Device>,
+        py::kw_only(),
+        py::arg("bias_tensor"),
+        py::arg("input_memory_config"),
+        py::arg("in_channels"),
+        py::arg("out_channels"),
+        py::arg("batch_size"),
+        py::arg("input_height"),
+        py::arg("input_width"),
+        py::arg("kernel_size"),
+        py::arg("stride"),
+        py::arg("padding"),
+        py::arg("dilation"),
+        py::arg("groups"),
+        py::arg("device"),
+        py::arg("conv_config") = std::nullopt);
+
+    module.def(
+        "prepare_conv_bias",
+        prepare_conv_bias<ttnn::MeshDevice>,
+        py::kw_only(),
+        py::arg("bias_tensor"),
+        py::arg("input_memory_config"),
+        py::arg("in_channels"),
+        py::arg("out_channels"),
+        py::arg("batch_size"),
+        py::arg("input_height"),
+        py::arg("input_width"),
+        py::arg("kernel_size"),
+        py::arg("stride"),
+        py::arg("padding"),
+        py::arg("dilation"),
+        py::arg("groups"),
+        py::arg("device"),
+        py::arg("conv_config") = std::nullopt);
+
+    module.def(
         "get_conv_padded_input_shape_and_mem_config",
         [](ttnn::Device* device,
            const ttnn::Tensor& input_tensor,
@@ -130,7 +212,7 @@ void py_bind_conv2d(py::module& module) {
            uint32_t width,
            uint32_t in_channels,
            uint32_t out_channels) -> std::tuple<ttnn::Shape, ttnn::MemoryConfig, bool, bool> {
-            return ttnn::operations::conv::conv2d::get_conv_padded_input_shape_and_mem_config<ttnn::Device>(
+            return get_conv_padded_input_shape_and_mem_config<ttnn::Device>(
                 device,
                 input_tensor,
                 conv_config,
@@ -160,7 +242,7 @@ void py_bind_conv2d(py::module& module) {
            uint32_t width,
            uint32_t in_channels,
            uint32_t out_channels) -> std::tuple<ttnn::Shape, ttnn::MemoryConfig, bool, bool> {
-            return ttnn::operations::conv::conv2d::get_conv_padded_input_shape_and_mem_config<MeshDevice>(
+            return get_conv_padded_input_shape_and_mem_config<MeshDevice>(
                 device,
                 input_tensor,
                 conv_config,
@@ -182,7 +264,7 @@ void py_bind_conv2d(py::module& module) {
 
     module.def(
         "convert_conv_weight_tensor_to_tiled_layout",
-        &ttnn::operations::conv::conv2d::convert_conv_weight_tensor_to_tiled_layout,
+        &convert_conv_weight_tensor_to_tiled_layout,
         py::arg("conv_weight_tensor").noconvert(),
         py::arg("in1_block_h"),
         py::arg("in1_block_w"),
@@ -190,7 +272,7 @@ void py_bind_conv2d(py::module& module) {
 
     module.def(
         "convert_conv_weight_tensor_to_special_padding_tiled_layout",
-        &ttnn::operations::conv::conv2d::convert_conv_weight_tensor_to_special_padding_tiled_layout,
+        &convert_conv_weight_tensor_to_special_padding_tiled_layout,
         py::arg("conv_weight_tensor").noconvert(),
         py::arg("in1_block_h"),
         py::arg("in1_block_w"),
@@ -198,7 +280,7 @@ void py_bind_conv2d(py::module& module) {
 
     module.def(
         "convert_conv_weight_tensor_to_grouped_layout",
-        &ttnn::operations::conv::conv2d::convert_conv_weight_tensor_to_grouped_layout,
+        &convert_conv_weight_tensor_to_grouped_layout,
         py::arg("conv_weight_tensor").noconvert(),
         py::arg("num_groups"),
         py::arg("output_dtype").noconvert() = std::nullopt);
@@ -214,7 +296,7 @@ void py_bind_conv2d(py::module& module) {
            const CoreCoord& compute_grid_size,
            ShardOrientation block_shard_orientation,
            bool is_out_tiled) -> ttnn::operations::sliding_window::ParallelConfig {
-            return ttnn::operations::conv::conv2d::determine_parallel_config(
+            return determine_parallel_config(
                 shard_layout, batch_size, input_channels, output_height, output_width, output_channels, compute_grid_size, block_shard_orientation, is_out_tiled);
         },
         py::arg("shard_layout"),
@@ -229,7 +311,7 @@ void py_bind_conv2d(py::module& module) {
 
     module.def(
         "create_sharded_memory_config_from_parallel_config",
-        &ttnn::operations::conv::conv2d::create_sharded_memory_config_from_parallel_config,
+        &create_sharded_memory_config_from_parallel_config,
         py::arg("tensor_shape"),
         py::arg("parallel_config"),
         py::arg("tile_size"));
