+++ ttnn/cpp/ttnn/operations/conv/conv2d/prepare_conv2d_weights.hpp
@@ -1,0 +1,65 @@
+
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+#include <optional>
+
+#include "conv2d_utils.hpp"
+#include "ttnn/core.hpp"
+#include "ttnn/operations/core/core.hpp"
+#include "ttnn/operations/matmul/matmul.hpp"
+#include "ttnn/operations/matmul/device/matmul_op.hpp"
+#include "ttnn/types.hpp"
+#include "ttnn/tensor/tensor_utils.hpp"
+#include "tt_metal/impl/dispatch/command_queue.hpp"
+#include "tt_metal/common/math.hpp"
+#include "ttnn/operations/data_movement/pad/pad.hpp"
+#include "ttnn/operations/conv/conv2d/device/conv2d_op.hpp"
+#include "ttnn/tensor/tensor.hpp"
+#include "ttnn/operations/sliding_window/sliding_window.hpp"
+#include "ttnn/operations/sliding_window/halo/halo.hpp"
+
+namespace ttnn {
+
+namespace operations::conv {
+namespace conv2d {
+template <typename T>
+ttnn::Tensor prepare_conv_weights(
+    const ttnn::Tensor& weight_tensor,
+    const ttnn::MemoryConfig& input_memory_config,
+    std::string weights_format,
+    uint32_t in_channels,
+    uint32_t out_channels,
+    uint32_t batch_size,
+    uint32_t input_height,
+    uint32_t input_width,
+    std::array<uint32_t, 2> kernel_size,
+    std::array<uint32_t, 2> stride,
+    std::array<uint32_t, 2> padding,
+    std::array<uint32_t, 2> dilation,
+    uint32_t groups,
+    T *device,
+    std::optional<const Conv2dConfig> conv_config_);
+
+template <typename T>
+ttnn::Tensor prepare_conv_bias(
+    const ttnn::Tensor& bias_tensor,
+    const ttnn::MemoryConfig& input_memory_config,
+    uint32_t in_channels,
+    uint32_t out_channels,
+    uint32_t batch_size,
+    uint32_t input_height,
+    uint32_t input_width,
+    std::array<uint32_t, 2> kernel_size,
+    std::array<uint32_t, 2> stride,
+    std::array<uint32_t, 2> padding,
+    std::array<uint32_t, 2> dilation,
+    uint32_t groups,
+    T *device,
+    std::optional<const Conv2dConfig> conv_config_);
+
+} // namespace conv2d
+} // namespace operations::conv
+} // namespace ttnn
