+++ ttnn/ttnn/__init__.py
@@ -296,7 +296,14 @@ from ttnn.operations.ccl import (
     Topology,
 )
 
-from ttnn.operations.conv2d import Conv2dConfig, get_conv_padded_input_shape_and_mem_config, get_conv_output_dim
+
+from ttnn.operations.conv2d import (
+    Conv2dConfig,
+    get_conv_output_dim,
+    get_conv_padded_input_shape_and_mem_config,
+    prepare_conv_weights,
+    prepare_conv_bias,
+)
 from ttnn.operations.pool import avg_pool2d
 from ttnn.operations.conv1d import Conv1d, Conv1dConfig
